hydra: refactor mpiexec/options.c
All the specific argument options are maintained in this file.
